Organized File Structure
Rerouted some pages and organized our components
