+++ flashcard/app/components/home_components/home_carousel.jsx
