+++ flashcard/app/components/home_components/home_footer.jsx
